secmet: split location functions into new file
