+++ antismash/common/secmet/locations.py
@@ -1,0 +1,62 @@
+# License: GNU Affero General Public License v3 or later
+# A copy of GNU AGPL v3 should have been included in this software package in LICENSE.txt.
+
+""" Helper functions for location operations """
+
+from typing import Tuple
+
+from Bio.SeqFeature import FeatureLocation, CompoundLocation
+
+
+def convert_protein_position_to_dna(start: int, end: int, location: FeatureLocation) -> Tuple[int, int]:
+    """ Convert a protein position to a nucleotide sequence position for use in generating
+        new FeatureLocations from existing FeatureLocations and/or CompoundLocations.
+
+        Arguments:
+            position: the position in question, must be contained by the location
+            location: the location of the related feature, for handling introns/split locations
+
+        Returns:
+            an int representing the calculated DNA location
+    """
+    if not 0 <= start < end <= len(location) // 3:
+        raise ValueError("Protein positions %d and %d must be contained by %s" % (start, end, location))
+    if location.strand == -1:
+        dna_start = location.start + len(location) - end * 3
+        dna_end = location.start + len(location) - start * 3
+    else:
+        dna_start = location.start + start * 3
+        dna_end = location.start + end * 3
+
+    # only CompoundLocations are complicated
+    if not isinstance(location, CompoundLocation):
+        if not location.start <= dna_start < dna_end <= location.end:
+            raise ValueError(("Converted coordinates %d..%d "
+                              "out of bounds for location %s") % (dna_start, dna_end, location))
+        return dna_start, dna_end
+
+    parts = sorted(location.parts, key=lambda x: x.start)
+    gap = 0
+    last_end = parts[0].start
+    start_found = False
+    end_found = False
+    for part in parts:
+        if start_found and end_found:
+            break
+        gap += part.start - last_end
+        if not start_found and dna_start + gap in part:
+            start_found = True
+            dna_start = dna_start + gap
+        if not end_found and dna_end + gap - 1 in part:
+            end_found = True
+            dna_end = dna_end + gap
+
+        last_end = part.end
+
+    assert start_found
+    assert end_found
+
+    if not location.start <= dna_start < dna_end <= location.end:
+        raise ValueError(("Converted coordinates %d..%d "
+                          "out of bounds for location %s") % (dna_start, dna_end, location))
+    return dna_start, dna_end
