+++ antismash/common/secmet/test/test_locations.py
@@ -1,0 +1,97 @@
+# License: GNU Affero General Public License v3 or later
+# A copy of GNU AGPL v3 should have been included in this software package in LICENSE.txt.
+
+# for test files, silence irrelevant and noisy pylint warnings
+# pylint: disable=no-self-use,protected-access,missing-docstring
+
+import unittest
+
+from antismash.common.secmet.locations import (
+    convert_protein_position_to_dna,
+    FeatureLocation,
+    CompoundLocation
+)
+
+
+class TestProteinPositionConversion(unittest.TestCase):
+    def setUp(self):
+        self.func = convert_protein_position_to_dna
+
+    def test_position_conversion_simple_forward(self):
+        location = FeatureLocation(0, 15, strand=1)
+        assert len(location) == 15
+        assert self.func(0, 2, location) == (0, 6)
+        assert self.func(1, 4, location) == (3, 12)
+
+    def test_position_conversion_simple_reverse(self):
+        location = FeatureLocation(0, 15, strand=-1)
+        assert len(location) == 15
+        assert self.func(0, 2, location) == (9, 15)
+        assert self.func(1, 4, location) == (3, 12)
+
+    def test_position_conversion_nonzero_start(self):
+        location = FeatureLocation(6, 21, strand=1)
+        assert len(location) == 15
+        assert self.func(0, 2, location) == (6, 12)
+        assert self.func(1, 4, location) == (9, 18)
+
+        location = FeatureLocation(6, 21, strand=-1)
+        assert len(location) == 15
+        assert self.func(0, 2, location) == (15, 21)
+        assert self.func(1, 4, location) == (9, 18)
+
+    def test_position_conversion_nonzero_start_compound(self):
+        location = CompoundLocation([FeatureLocation(6, 18, strand=1),
+                                     FeatureLocation(24, 27, strand=1)])
+        assert len(location) == 15
+        assert self.func(0, 2, location) == (6, 12)
+        assert self.func(1, 4, location) == (9, 18)
+        assert self.func(3, 5, location) == (15, 27)
+
+        location = CompoundLocation([FeatureLocation(6, 15, strand=-1),
+                                     FeatureLocation(21, 27, strand=-1)])
+        assert len(location) == 15
+        assert self.func(0, 2, location) == (21, 27)
+        assert self.func(1, 4, location) == (9, 24)
+        assert self.func(3, 5, location) == (6, 12)
+
+    def test_position_conversion_compound_forward(self):
+        location = CompoundLocation([FeatureLocation(0, 6, strand=1),
+                                     FeatureLocation(9, 18, strand=1)])
+        assert len(location) == 15
+        assert self.func(0, 4, location) == (0, 15)
+        assert self.func(1, 5, location) == (3, 18)
+
+        location = CompoundLocation([FeatureLocation(0, 6, strand=1),
+                                     FeatureLocation(12, 15, strand=1),
+                                     FeatureLocation(21, 27, strand=1)])
+        assert len(location) == 15
+        assert self.func(0, 4, location) == (0, 24)
+        assert self.func(1, 5, location) == (3, 27)
+        assert self.func(2, 3, location) == (12, 15)
+
+    def test_position_conversion_compound_reverse(self):
+        location = CompoundLocation([FeatureLocation(0, 6, strand=-1),
+                                     FeatureLocation(9, 18, strand=-1)])
+        assert len(location) == 15
+        assert self.func(0, 4, location) == (3, 18)
+        assert self.func(1, 5, location) == (0, 15)
+
+        location = CompoundLocation([FeatureLocation(0, 6, strand=-1),
+                                     FeatureLocation(12, 15, strand=-1),
+                                     FeatureLocation(21, 27, strand=-1)])
+        assert len(location) == 15
+        assert self.func(0, 4, location) == (3, 27)
+        assert self.func(1, 5, location) == (0, 24)
+        assert self.func(2, 3, location) == (12, 15)
+
+    def test_other(self):
+        location = CompoundLocation([FeatureLocation(5922, 6190, strand=1),
+                                     FeatureLocation(5741, 5877, strand=1),
+                                     FeatureLocation(4952, 5682, strand=1)])
+        assert self.func(97, 336, location) == (5243, 6064)
+
+        location = CompoundLocation([FeatureLocation(5922, 6190, strand=-1),
+                                     FeatureLocation(5741, 5877, strand=-1),
+                                     FeatureLocation(4952, 5682, strand=-1)])
+        assert self.func(97, 336, location) == (5078, 5854)
